+++ features/style/api/get-styles.ts
@@ -5,8 +5,9 @@ import type { Style } from '../types';
 
 /**
  * 指定したItemのスタイル一覧を取得
+ * itemId が null の場合はグローバルスタイルを取得
  */
-export async function getStyles(itemId: string): Promise<Style[]> {
+export async function getStyles(itemId: string | null): Promise<Style[]> {
   const styles = await prisma.style.findMany({
     where: { itemId },
     orderBy: { order: 'asc' },
