+++ features/style/index.ts
@@ -1,2 +1,3 @@
 export * from './api';
+export * from './components';
 export * from './types';
